Refactor: conditional rendering added for Post when not logged in
