+++ src/components/pages/Post/Post.module.css
@@ -1,0 +1,7 @@
+.container {
+    display: flex;
+    justify-content: center;
+align-items: center;
+height: calc(100vh - 80px);
+min-height: 520px;
+}
